+++ src/guard/isConsentAttributes.ts
@@ -17,9 +17,41 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-import { CONSENT_STATUS } from "../constants";
+import { UrlString } from "@inrupt/solid-client";
+import { CONSENT_STATUS, RESOURCE_ACCESS_MODE } from "../constants";
+import { ConsentAttributes } from "../type/AccessVerifiableCredential";
+import { ResourceAccessMode } from "../type/ResourceAccessMode";
+import { isUnknownObject } from "./isUnknownObject";
 import { ConsentStatus } from "../type/ConsentStatus";
 
-export function isConsentStatus(x: unknown): x is ConsentStatus {
+function isResourceAccessModeArray(x: unknown): x is Array<ResourceAccessMode> {
+  return (
+    Array.isArray(x) &&
+    x
+      .map((y) => RESOURCE_ACCESS_MODE.has(y))
+      .reduce((previous, current) => previous && current)
+  );
+}
+
+function isConsentStatus(x: unknown): x is ConsentStatus {
   return typeof x === "string" && (CONSENT_STATUS as Set<string>).has(x);
+}
+
+// TODO: Discuss a strongly typed UrlString guard (as a team).
+function isStringArray(x: unknown): x is Array<UrlString> {
+  return (
+    Array.isArray(x) &&
+    x
+      .map((y) => typeof y === "string")
+      .reduce((previous, current) => previous && current)
+  );
+}
+
+export function isConsentAttributes(x: unknown): x is ConsentAttributes {
+  return (
+    isUnknownObject(x) &&
+    isResourceAccessModeArray(x.mode) &&
+    isConsentStatus(x.hasStatus) &&
+    isStringArray(x.forPersonalData)
+  );
 }